Fix and use readerIsEmpty()
The function originally checked if the buffer was NOT FULL, which is not
the same as being empty. Now, READER_EMP bit is checked in the reader's
flags.

Further, the function was not being utilized. Utility functions like
this should always be used where applicable since it reduces the chances
of creating bugs.